requirements.txt
asgiref==3.8.1
Django==5.0.6
djangorestframework==3.15.1
pyaes==1.6.1
pyasn1==0.6.0
rsa==4.9
sqlparse==0.5.0
Telethon==1.23.0
